prevent players from leaving jail (wyvern no geo ..) by vital cdl and buylist fix by striker - big thanks !
git-svn-id: http://svn.l2jfree.com/l2jfree@4740 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/clientpackets/RequestBuyItem.java
@@ -261,7 +261,7 @@ public class RequestBuyItem extends L2GameClientPacket
 				return;
 			}
 
-			long stackPrice = (price * count);
+			long stackPrice = (long)price * count;
 			long taxedPrice = (long) (stackPrice * taxRate);
 			long tax = taxedPrice - stackPrice;
 			if (taxedPrice >= Integer.MAX_VALUE)
